refactor: 아고라 생성 버튼 로직 분리
