+++ src/app/(main)/_lib/postCreateAgora.ts
@@ -1,4 +1,4 @@
-import { AgoraConfig, AgoraId } from '@/app/model/Agora';
+import { AgoraConfig, CreateAgoraResponse } from '@/app/model/Agora';
 import { callFetchWrapper } from '@/lib/fetchWrapper';
 import { getSession } from '@/serverActions/auth';
 import { AUTH_MESSAGE, SIGNIN_REQUIRED } from '@/constants/authErrorMessage';
@@ -15,10 +15,6 @@ const CAPACITY_NULL = { categoryId: '널이어서는 안됩니다' };
 const DURATION_UNDER_ERROR = { duration: '1 이상이어야 합니다' };
 const DURATION_OVER_ERROR = { duration: '180 이하이어야 합니다' };
 const FILE_SIZE_OVER_ERROR = 'File size cannot exceed 5MB.';
-
-type CreateAgoraResponse = {
-  id: AgoraId;
-};
 
 export const postCreateAgora = async (
   info: AgoraConfig,
